move delete handler into pathery
