+++ packages/pathery/src/service/mod.rs
@@ -1,1 +1,2 @@
+pub mod doc;
 pub mod index;
